move static class functions into a helper file
