+++ trekpedia/trekpedia.py
@@ -1,13 +1,20 @@
 """Define the Trekpedia class."""
 from __future__ import annotations
 
-import json
 import re
 from typing import TYPE_CHECKING, Dict, List, Tuple
 
 import requests
 from bs4 import BeautifulSoup
-from rich import print  # pylint: disable=redefined-builtin
+from rich import print
+
+from trekpedia.helpers import (  # pylint: disable=redefined-builtin
+    clean_string,
+    get_overview_rows,
+    parse_url,
+    print_season_header,
+    save_json,
+)
 
 if TYPE_CHECKING:
     from requests import Response
@@ -35,7 +42,7 @@ class Trekpedia:
 
     def get_summary_data(self) -> None:
         """Get and parse the summary data."""
-        self.series_markup: BeautifulSoup = self.parse_url(self.main_url)
+        self.series_markup: BeautifulSoup = parse_url(self.main_url)
 
     def get_series_detail_link(self, url: str) -> str:
         """Return the link to the detail page for the specified series."""
@@ -142,7 +149,7 @@ class Trekpedia:
         # tas, ds9, voy etc) just skip over this one as it is a
         # summary...
         try:
-            episode_data["num_overall"] = self.clean_string(
+            episode_data["num_overall"] = clean_string(
                 episode.find("th").text, brackets=True
             )
         except AttributeError:
@@ -201,7 +208,7 @@ class Trekpedia:
         # get the required data using the header indexes, otherwise
         # will mess up on ds9-s4 and later since they add new
         # columns to the table.
-        episode_data["title"] = self.clean_string(
+        episode_data["title"] = clean_string(
             cells[headers.index("title")].text.replace('"', ""),
             brackets=True,
         )
@@ -219,7 +226,7 @@ class Trekpedia:
                 # set the link url to an empty string...
                 episode_data["link"] = ""
 
-        episode_data["director"] = self.clean_string(
+        episode_data["director"] = clean_string(
             cells[headers.index("directed_by")].text, brackets=True
         )
 
@@ -232,11 +239,11 @@ class Trekpedia:
             or re.search("^paramount.*date$", item)
         ][0]
         try:
-            episode_data["air_date"] = self.clean_string(
+            episode_data["air_date"] = clean_string(
                 cells[airdate_idx].text, brackets=True
             )
         except IndexError:
-            episode_data["air_date"] = self.clean_string(
+            episode_data["air_date"] = clean_string(
                 cells[-1].text, brackets=True
             )
 
@@ -258,7 +265,7 @@ class Trekpedia:
         # series and even seasons! at this time we also remove any
         # unicode stuff
         headers = [
-            self.clean_string(
+            clean_string(
                 x.text, underscores=True, brackets=True, lowercase=True
             )
             for x in table[0].find_all("th")
@@ -290,11 +297,11 @@ class Trekpedia:
 
         filename = self.get_json_filename(index, series)
 
-        self.print_season_header(series, filename)
+        print_season_header(series, filename)
 
         season_all = {}
 
-        self.episode_markup = self.parse_url(series["episodes_url"])
+        self.episode_markup = parse_url(series["episodes_url"])
 
         try:
             overview_table = self.episode_markup.find(
@@ -316,7 +323,7 @@ class Trekpedia:
                 }
             else:
                 # its a standard Series with multiple seasons.
-                for season in self.get_overview_rows(overview_table):
+                for season in get_overview_rows(overview_table):
                     try:
                         overview_row_header = season.find("th")
                         overview_row_data = season.find_all("td")
@@ -354,7 +361,7 @@ class Trekpedia:
                 f"({err}) at line number: {err.__traceback__.tb_lineno}"
             )
             return
-        self.save_json(filename, {"seasons": season_all})
+        save_json(filename, {"seasons": season_all})
 
     def consolidate(self, series, season_number, episodes, overview_row_data):
         """Consolidate the season data into a dict ready for JSON."""
@@ -365,11 +372,11 @@ class Trekpedia:
         except ValueError:
             offset = 0
 
-        season_start = self.clean_string(
+        season_start = clean_string(
             " ".join(overview_row_data[1 + offset].text.split()),
             brackets=True,
         )
-        season_end = self.clean_string(
+        season_end = clean_string(
             " ".join(overview_row_data[2 + offset].text.split()),
             brackets=True,
         )
@@ -379,9 +386,7 @@ class Trekpedia:
         if series["name"].lower() == "discovery" and season_number == 1:
             season_end = "February 11, 2018"
         return {
-            "total": self.clean_string(
-                overview_row_data[0].text, brackets=True
-            ),
+            "total": clean_string(overview_row_data[0].text, brackets=True),
             "season_start": season_start,
             "season_end": season_end,
             "episodes": episodes,
@@ -394,48 +399,6 @@ class Trekpedia:
         )
 
         return filename
-
-    @staticmethod
-    def print_season_header(series, filename) -> None:
-        """Display the header for each series as it is processed."""
-        print(f'Processing : [bold][underline]{series["name"]}')
-        print(f"  -> Using URL : [green]{series['episodes_url']}")
-        print(f"  -> Storing episodes to [green]'{filename}'")
-
-    @staticmethod
-    def get_overview_rows(summary_table):
-        """Return markup for the rows in the series overview table."""
-        summary_rows = summary_table.find("tbody").find_all("tr")[2:]
-        return summary_rows
-
-    @staticmethod
-    def save_json(filename, data):
-        """Save the specified data as a JSON file to the specified location."""
-        with open(filename, "w", encoding="utf-8") as file:
-            json.dump(data, file, ensure_ascii=False, indent=4)
-
-    @staticmethod
-    def clean_string(
-        dirty_string: str, underscores=False, brackets=False, lowercase=False
-    ) -> str:
-        """Take a string and remove underscores, spaces etc as required."""
-        if underscores:
-            dirty_string = (
-                dirty_string.replace(" ", "_")
-                .replace(".", "_")
-                .replace("__", "_")
-            )
-        if brackets:
-            dirty_string = "".join(re.split(r"[\(\)\[\]]", dirty_string)[::2])
-        if lowercase:
-            dirty_string = dirty_string.lower()
-        return " ".join(dirty_string.split())
-
-    @staticmethod
-    def parse_url(url) -> BeautifulSoup:
-        """Get the specified url and parse with BeautifulSoup."""
-        result: Response = requests.get(url, timeout=10)
-        return BeautifulSoup(result.text, "lxml")
 
 
 if __name__ == "__main__":
